extract common collection methods into separate class
