+++ src/Core/Grid/Action/BulkActionCollection.php
@@ -26,6 +26,8 @@
 
 namespace PrestaShop\PrestaShop\Core\Grid\Action;
 
+use PrestaShop\PrestaShop\Core\Grid\Collection\AbstractCollection;
+
 final class BulkActionCollection extends AbstractCollection implements BulkActionCollectionInterface
 {
     /**
